remove the double api call from graph building
Closes #478
