+++ app/javascript/packs/application.js
@@ -42,7 +42,6 @@ import "../graph_builder.js"
 import "../run_claim.js"
 import "../run_edit.js"
 import "../search.js"
-import "../segment_graph_builder.js"
 import "../timeline.js"
 import "../tooltips.js"
 import "../twitch.js"
